requirements.txt
asgiref==3.7.2
Django>=3.2.6
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==20.1.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
PyJWT==1.7.1